stages: add new symlink stage
This introduces a new stage, `org.osbuild.symlink`, for creating
symbolic links within the filesystem tree.

This takes a list of paths, each with a `source` and a `link`
parameters, inspired by the copy stage.

The source is a simple string because we want the link to be either
absolute or relative.